moved pytorch model into descriptors folder
